// src/lib/olxApiClient/api/schema.ts
export const KeySchema = z.enum([
  "builttype",
  "floor_select",
  "furniture",
  "m",
  "price",
  "price_per_m",
  "market",
  "rent",
  "rooms",
  "roomsize",
  "preferences",
]);
export type Key = z.infer<typeof KeySchema>;

export const CityNameSchema = z.string();
export type CityName = z.infer<typeof CityNameSchema>;

export const NormalizedNameSchema = z.string();
export type NormalizedName = z.infer<typeof NormalizedNameSchema>;

export const OfferTypeSchema = z.enum(["offer"]);
// ...
export const DatumStatusSchema = z.enum([
  "active",
  "removed_by_user",
  "disabled",
  "removed_by_moderator",
]);
// ...
export const UserSchema = z.object({
  id: z.number(),
  created: z.string(),
  other_ads_enabled: z.boolean(),
  name: z.string(),
  logo: z.union([z.null(), z.string()]),
  photo: z.union([z.null(), z.string()]),
  banner_mobile: z.string(),
  banner_desktop: z.string(),
  company_name: CompanyNameSchema,
  about: z.string(),
  b2c_business_page: z.boolean(),
  is_online: z.boolean(),
  last_seen: z.string(),
  seller_type: z.null(),
  uuid: z.string(),
});
// ...
export const PhotoSchema = z.object({
  id: z.number(),
  filename: z.string(),
  rotation: z.number(),
  width: z.number(),
  height: z.number(),
  link: z.string(),
});
// ...
export const ParamSchema = z.discriminatedUnion("key", [
  AreaParam,
  RentParam,
  PriceParam,
  FurnitureParam,
  RoomsParam,
  BuiltTypeParam,
  FloorParam,
  RoomSizeParam,
  PreferencesParam,
  PricePerM,
  Market,
]);
export type Param = z.infer<typeof ParamSchema>;

export const MapSchema = z.object({
  zoom: z.number(),
  lat: z.number(),
  lon: z.number(),
  radius: z.number(),
  show_detailed: z.boolean(),
});
export type Map = z.infer<typeof MapSchema>;

export const DistrictSchema = z.object({
  id: z.number(),
  name: z.string(),
});
export type District = z.infer<typeof DistrictSchema>;

export const CitySchema = z.object({
  id: z.number(),
  name: CityNameSchema,
  normalized_name: NormalizedNameSchema,
});
export type City = z.infer<typeof CitySchema>;

export const LocationSchema = z.object({
  city: CitySchema,
  district: z.union([DistrictSchema, z.null()]).optional(),
  region: CitySchema,
});
export type Location = z.infer<typeof LocationSchema>;

export const RockSchema = z.object({
  offer_id: z.null(),
  active: z.boolean(),
  mode: ModeSchema,
});
export type Rock = z.infer<typeof RockSchema>;

export const DeliverySchema = z.object({
  rock: RockSchema,
});
export type Delivery = z.infer<typeof DeliverySchema>;

export const ContactSchema = z.object({
  name: z.string(),
  phone: z.boolean(),
  chat: z.boolean(),
  negotiation: z.boolean(),
  courier: z.boolean(),
});
export type Contact = z.infer<typeof ContactSchema>;

export const CategorySchema = z.object({
  id: z.number(),
  type: CategoryTypeSchema,
});
export type Category = z.infer<typeof CategorySchema>;

export const DatumSchema = z.object({
  id: z.number(),
  url: z.string(),
  title: z.string(),
  last_refresh_time: z.string(),
  created_time: z.string(),
  valid_to_time: z.string(),
  pushup_time: z.union([z.null(), z.string()]),
  description: z.string(),
  params: z.array(ParamSchema),
  key_params: z.array(KeySchema),
  business: z.boolean(),
  user: UserSchema,
  status: DatumStatusSchema,
  contact: ContactSchema,
  map: MapSchema,
  location: LocationSchema,
  photos: z.array(PhotoSchema),
  category: CategorySchema,
  offer_type: OfferTypeSchema,
  external_url: z.union([z.null(), z.string()]).optional(),
});
export type Datum = z.infer<typeof DatumSchema>;

export const OlxApiResponseSchema = z.object({
  data: z.array(DatumSchema),
});